+++ app/Http/Controllers/Network/InternetController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Network;
 
 use App\Models\EventTrigger;
 use App\Models\InternetAccess;
@@ -9,6 +9,8 @@ use App\Models\Role;
 use App\Models\User;
 use App\Models\WifiConnection;
 use App\Utils\TabulatorPaginator;
+use App\Http\Controllers\Controller;
+
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
