+++ app/Http/Controllers/Secretariat/PermissionController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Secretariat;
 
 use App\Http\Controllers\Controller;
 use App\Models\Role;
