+++ app/Http/Controllers/Secretariat/SemesterController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers\Secretariat;
 
 use App\Http\Controllers\Controller;
 use App\Models\User;
